+++ docs/src/reference/cli/changelog.rst
@@ -13,6 +13,16 @@ Added
 
 * Error out when a user tries to run a Dataflow-based job with Stackdriver log-based metrics client configured.
 
+Fixed
+*****
+
+* Correctly validate existence of Dataflow-related Klio config when running on Dataflow (and not just "not --direct-runner").
+
+Changed
+*******
+
+* When running a job, effective config is no longer written to ``klio-job-run-effective.yaml``, but instead to a temp file.  This file no longer needs to be included in ``setup.py`` projects (See `PR 233 <https://github.com/spotify/klio/pull/233>`_).
+
 .. end-21.10.0
 
 .. _cli-21.9.0: